Update patch set 2
Patch Set 2:

Hello,

The "netperf" based networking tests were not part of the official list of networking tests. 

You can find the network tests included for B release in the link below
https://wiki.opnfv.org/display/qtip/Platform+Performance+Benchmarking (Network Test cases in the above link)

We tried and developed code for "netperf" (and you can find the ansible playbook within the benchmarks/playbooks directory)
But we didn't include this officially for the B release. We found that "ipef" was a better option to use as compared to "netperf". Additionally, "netperf" was no longer maintained by the upstream developers  and was comparatively a very old tool. 
Hence it was not included in the network test list.

Patch-set: 2
Reviewer: Gerrit User 837 <837@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0