Update patch set 1
Patch Set 1:

(3 comments)

Not going to +1 or -1 this, as it looks like the real work here is just changing from relative path to pull path with a variable.

There are definitely some things that could be done better in the Ansible plays though, and you might want to take note of that for future enhancements.

Patch-set: 1
Reviewer: Gerrit User 3299 <3299@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0