Update patch set 10
Patch Set 10:

Tried to resume the original patch for review.

It enables user to run a single test case. However, the original feature of running a suite seems to be no longer supported. This is not an expected result.

We should extend the existing function instead of replacing it. Running a suite is still useful in some condition.

@Taseer, could you modify the patch to keep the possibility to run a suite while adding the new choice?

Patch-set: 10