Update patch set 1
Patch Set 1: Code-Review+1

> > I think the name we agreed on is `suite`
 > 
 > I just think test_suite is better.

The samples (compute, network and storage) are close to a suite of test. But what actually we are targeting is application in our vision.

For now, it is ok to name it as test_suite. But we might need a better name other than test_suite or suite.

Patch-set: 1
Label: Code-Review=+1