Update patch set 1
Patch Set 1:

(1 comment)

Would you give a summary reason of this bug in commit message?

It seems to be caused by missing `.yaml` in calling `get_index`. If `.yaml` is the only format supported, we may auto-complete it in `get_index` instead of modifying arguments in each calling.

Patch-set: 1
Label: Code-Review=0