Update patch set 3
Patch Set 3:

I have just realized there is a problem with this implementation.

The description has no "actual" link with the available suites. When we add a new one in the `benchmarks/suite` folder, the result is not updated automatically.

For now, the suite file does not include any meta data for display. We may simply list the file names and generate an auto description such as "compute benchmark suite".

Next step would be modifying the format of suite file to embed the meta data.

Excuse me for not pointing it out earlier.

Patch-set: 3