fix testclassname  and file name support for both type CheckXxx and TestXxx
rename test classes and files to CheckXxx and check_xxx [QTIP-182]. Many other
classes in other directories also have test files with `test` in their file and classes name. So need to decide whether to make `Check` as prefix as a standard or not.

rename TestXxx to CheckXxx for all three test files
JIRA: QTIP-182

Signed-off-by: Akhil Batra <akhil.batra@research.iiit.ac.in>
Change-Id: I71f0ca3a5406da64b6ab00ff54151dbc6baa76ae