Update patch set 2
Patch Set 2: Code-Review-1

Personally I prefer the original naming, which is with prefix 'test_'. tox is not always introduced to run the unittest. Further more if we want to run part of the tests rather than all, we must use pytest directly, in this condition, if we change the naming rule, it will take us extra work to do the configuration, which make it quite user-unfriendly. And, not all the people use tox.

Patch-set: 2
Reviewer: Gerrit User 2651 <2651@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1