Update patch set 2
Patch Set 2:

(3 comments)

To keep you on the same page, I had designed it for viewing the LOGS or JSON file of the Results. 

Let me know if you have other plans. Like graphs of specified benchmarks etc. Graphics are not supported in docker containers, but since we are making qtip independant......

Patch-set: 2