Update patch set 1
Patch Set 1:

`describe()` was designed for this purpose. We choose this name because it provide information only.

The information could be `show` in CLI or web page and also possible to be used for other purpose, for example, dumping a report. So `show` is a method in CLI, not in runners

Patch-set: 1