Update patch set 1
Patch Set 1:

> > We shall no longer have `perftest`.
 > >
 > > Please note the changes in https://gerrit.opnfv.org/gerrit/#/c/26309/
 > > , rebase to it and use `/tests/data/benchmarks` for unit test.
 > 
 > No perftest ? Any particular reason ? Would it just be Suite ?

Please refer to https://wiki.opnfv.org/display/qtip/Design perftest is renamed to `metric` and referred in `case`. Now we shall have `plan`, `suite` and `case` runner. But `suite` and `case` definition shall be embedded in `plan` since it is unlikely to be reused. The reusable assets would be `QPI` spec and `metric` spec.

Patch-set: 1