Add aarch64 patch
This commit adds a patch that can be applied to the Dockerfile to adapt
it for aarch64. When applied from releng, a new Dockerfile.aarch64 is
created and an aarch64 qtip image is created.

The aarch64 qtip image will be build with the releng fix
https://gerrit.opnfv.org/gerrit/30181

JIRA: ARMBAND-219

Change-Id: I551ae7fa1d7fc23c6e5f6381d605b0b22dbec695
Signed-off-by: Cristina Pauna <cristina.pauna@enea.com>
(cherry picked from commit 2935006a650f92255c8046d236e691360148d9ff)