Update patch set 3
Patch Set 3: Code-Review+1

> sorry I test it in virtual env and non-virtual env, both of them
 > don't work, I think the problem happens with the installation
 > config, non-python-files are not carried together, I will submit a
 > patch to fix it

Do you want to submit another patch or wait for Taseer to amend this?

We know the path currently is a problem. The goal is to make it work in unit test and also when installed in a new environment.

Patch-set: 3
Reviewer: Gerrit User 4469 <4469@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=+1