Update patch set 2
Patch Set 2:

> > (1 comment)
 > 
 > Is it something critical ?
 > If yes then kindly let me know a better way to do it.

I think @Serena may probably worried about you need to maintain the test case frequently when the templates are modified.

We will move `templates` to `/qtip/benchmarks/metric` so that it won't be part of qtip **code** distribution. 

What I suggest for reporter testing is try to feed it with invalid or incomplete data to see if exceptions are well handled and user got proper help message to understand what is going wrong. That is something we need to enhance in reporter.

Patch-set: 2