Update patch set 1
Patch Set 1:

> The run part executes correctly when I enter eval $(ssh-agent) from
 > terminal directly, but not when I have it done via the playbook.

I think ansible execute shell command in a child process. And the environment variable is not exported to current process. One solution would be capture the output of `ssh-agent` and set environment variable manually from ansible tasks. Not sure if there is any way easier to do it.

Patch-set: 1