Update patch set 3
Patch Set 3: -Code-Review

Similar question as in https://gerrit.opnfv.org/gerrit/#/c/37577/ but in this case the log file seems more complex and I think it could be better if it is a structured format such as json.

Think one step further, I tend to separate the collecting of **original** logs and configuration file from parsing the content. The latter could be done in a local python script and easier to verify.

Patch-set: 3