Update patch set 2
Patch Set 2:

I'm OK with the interface, but have some doubts on the plugin management. It seems animate-css, boostrap, jquery-* are all dependent plugins. Can we manage them with a frontend package manager, e.g. bower (or something else)? So that we don't need to include them in the repository.

Patch-set: 2