Update patch set 2
Patch Set 2:

(1 comment)

> I hope we would be following the same framework for storperf as
 > well ?

The test running is more or less the same. @Zhihui has ported the testing script for local use in qtip. The main difference is that we don't need to build storperf images but pull directly from docker store.

In addition, we need to collect system information with ceph-* roles, do calculation and generate report.

Patch-set: 2
Label: Verified=0