Update patch set 4
Patch Set 4:

> (1 comment)
 > 
 > > > what will be snaps used for?
 > >
 > > For launching VMs or a Stack. Using a project that already does
 > > this, rather than writing from scratch
 > 
 > Please consider how to integrate it into current QTIP framework
 > before going deep.
 > 
 > It seems to be a Python library wrapping OpenStack operations and
 > suitable for integrating in a service oriented framework. But I'm
 > not sure how we could make it work with command oriented framework
 > like Ansible which is the current reality of QTIP.

Yes, you are right about it being a mapping. What I was thinking that the `setup` command create a stack, and generate an inventory of the floating IP. That way we can use the existing QTIP framework.

I will try and do it in this same patch, to get a more clear picture.

Patch-set: 4